Pendulumクラスの分離
