+++ main.cpp
@@ -1,4 +1,5 @@
 #include <SDL.h>
+#include <pendulum.h>
 
 #include <cmath>
 #include <iostream>
@@ -19,51 +20,6 @@ bool load_media();
 void close();
 
 const double PI = 3.14159265359;
-
-class Pendulum
-{
-    /*
-    振り子を表すクラス.
-    振り子の座標は, 左の方向にx軸, 下の方向にy軸が伸びている.
-    振り子の角度は, x軸の正の向きとの成す角.
-    */
-private:
-    int pendulum_num;               // 振り子の数
-    double pendulum_string_length;  // 振り子の間の糸の長さ
-    // vector<pair<double, double>> pendulum_coords; //
-    // それぞれの振り子の中心の座標
-    vector<double>
-        pendulum_thetas;  // それぞれの振り子の中心とx軸の正の向きとの成す角
-
-public:
-    Pendulum(int pendulum_num, double pendulum_string_length)
-        : pendulum_num(pendulum_num),
-          pendulum_string_length(pendulum_string_length)
-    {
-        // 下方向に振り子が伸びるよう初期化
-        // pendulum_coords = vector<pair<double, double>>(pendulum_num);
-        pendulum_thetas = vector<double>(pendulum_num, PI / 2.0);
-        // for(int i = 0; i < pendulum_num; i++) {
-        //     pendulum_coords[i] = {0.0,
-        //                           pendulum_string_length * (double)(i + 1)};
-        // }
-    }
-
-    vector<pair<double, double>> compute_coords() const
-    {
-        /*固定端を原点とする振り子の座標を返す.*/
-        vector<pair<double, double>> ret(pendulum_num);
-        double now_x = 0.0, now_y = 0.0;  // 1つ上の振り子の座標
-        for(int i = 0; i < pendulum_num; i++) {
-            now_x = now_x + pendulum_string_length * cos(pendulum_thetas[i]);
-            now_y = now_y + pendulum_string_length * sin(pendulum_thetas[i]);
-            ret[i] = {now_x, now_y};
-        }
-        return ret;
-    }
-
-    int get_pendulum_num() const { return pendulum_num; }
-};
 
 // SDLを初期化する
 // 返り値:成功したか否か
